fix(core): rename folder.
