+++ angularExpensesTracker/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthService } from './shared/auth.service';
+import { AuthService } from './core/auth.service';
 import {take} from 'rxjs/operators';
