+++ angularExpensesTracker/src/app/login/login.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 
 import { take } from 'rxjs/operators';
 
-import { AuthService } from '../shared/auth.service';
+import { AuthService } from '../core/auth.service';
 
 interface Login {
   username: string;
